+++ src/Exceptions/BadPropertyAccessException.php
@@ -1,0 +1,26 @@
+<?php
+declare(strict_types=1);
+
+namespace Stancer\Exceptions;
+
+use Stancer\Interfaces\ExceptionInterface;
+use Psr;
+
+/**
+ * Exception thrown if a callback refers to an undefined method or if some arguments are missing.
+ */
+class BadPropertyAccessException extends Exception implements ExceptionInterface
+{
+    /** @var string Default log level */
+    protected static $logLevel = Psr\Log\LogLevel::CRITICAL;
+
+    /**
+     * Return default message for that kind of exception.
+     *
+     * @return string
+     */
+    public static function getDefaultMessage(): string
+    {
+        return 'Bad property access';
+    }
+}